api_yamdb/requirements.txt
asgiref==3.6.0
Django==3.2
django-filter
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
Pillow
djoser
requests==2.26.0